Migrate passwordsPrivate.isAccountStorageEnabled to a Mojo interface
Bug: 432409279
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7201029
Commit-Queue: Andrii Natiahlyi <natiahlyi@google.com>
Cr-Commit-Position: refs/heads/main@{#1550447}

+++ chrome/browser/resources/password_manager/password_manager_proxy.ts
@@ -696,7 +696,11 @@ export class PasswordManagerImpl implements PasswordManagerProxy {
   }
 
   isAccountStorageEnabled() {
-    return chrome.passwordsPrivate.isAccountStorageEnabled();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      return chrome.passwordsPrivate.isAccountStorageEnabled();
+    }
+    return this.handler.isAccountStorageEnabled().then(
+        result => result.enabled);
   }
 
   setAccountStorageEnabled(enabled: boolean) {
@@ -732,10 +736,11 @@ export class PasswordManagerImpl implements PasswordManagerProxy {
   }
 
   deleteAllPasswordManagerData() {
-    return loadTimeData.getBoolean('enablePasswordManagerMojoApi') ?
-        this.handler.deleteAllPasswordManagerData().then(
-            result => result.success) :
-        chrome.passwordsPrivate.deleteAllPasswordManagerData();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      return chrome.passwordsPrivate.deleteAllPasswordManagerData();
+    }
+    return this.handler.deleteAllPasswordManagerData().then(
+        result => result.success);
   }
 
   getActorLoginPermissions() {
